sai_stores/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-jazzmin==3.0.1
django-static==1.5.6
django-tinymce==4.1.0
idna==3.10
pillow==11.1.0
razorpay==1.4.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
